feat(historyBasketProducts): receiving list of products

+++ Assignment_#1/src/app/src/main/java/com/cpm/g1/theacmeelectronicsshop/ui/basket/CheckoutActivity.kt
@@ -45,7 +45,7 @@ class CheckoutActivity: AppCompatActivity() {
 
         cursor.moveToFirst()
         while(!cursor.isAfterLast) {
-            products.add(Product(cursor.getInt(0), cursor.getInt(1)))
+            products.add(Product(cursor.getLong(0), cursor.getInt(1)))
             cursor.moveToNext()
         }
         cursor.close()
